``` Fix: Correct volume fixing factor calculation (master)
The volume fixing factor was incorrectly calculated as the
cube root of the ratio of sphere volume to initial
volume. This commit corrects the calculation to be the
direct ratio of sphere volume to initial volume.  Also
added print statements for initial and sphere volumes and
added radius fixing factor to the volume info box.
```