``` feat: Implement center of mass alignment (master)
Calculates and applies center of mass alignment to ensure proper
orientation of the nuclear shape during fission.  This uses
numerical integration to determine the z-coordinate of the
center of mass, assuming constant nuclear density.  The mass
element is proportional to the volume element dV = πρ²dz.

The calculated z_cm is then used to shift the shape along
the z-axis, aligning it correctly for visualization.  The
scaling factor 'c' is currently set to 1.
```

Signed-off-by: Aleksander Augustyn <aleksander.augustyn@ncbj.gov.pl>